corrected variable type in declaration

+++ 0x07-pointers_arrays_strings/1-memcpy.c
@@ -10,7 +10,7 @@
 
 char *_memcpy(char *dest, char *src, unsigned int n)
 {
-	size_t = a;
+	unsigned int = a;
 
 	for (a = 0; a < n; a++)
 		dest[a] = src[a];
